Integrate data into canvas
Data can now be integrated into canvas, the scale and transformIndex
methods take care of translating canvas indices into price values.
data.js has also been improved to save its keys as string, this way,
users can have more control with regards to the decimal lengths of
their charts. Data.js can also take any tickSize value as it will merge
the limit order quantities accordingly. It calculates the nearest
tick value on the scale for appending the limit orders. This may show
a different result when comparing it to Binance itself, however, when
comparing the most granular ticksize (for BTCUSDT it is 0.10) it was
observed to be highly accurate.