Add new canvas for rect fills, move eventdispatch
Realistically, calculating price buckets should be done on the
backend because it requires summing a lot of price points at a
certain period. I moved the event dispatch to end of process to
simulate this, and to make sure the vis data is updated after the
operations, the event parameter, however, was passed through multiple
functions which I don't like. I can change this next time, but might
stick with it for now